requirements.txt
requests
python-dotenv
PyYAML
openai
python-gitlab
python-dateutil
Jinja2